SCRUM-12 refactor: вынесено модальное окно создания формы в переиспользуемый компонент
