+++ src/features/career-form/ui/career-forms/hooks/index.ts
@@ -1,2 +1,1 @@
 export { useCareerFormActions } from './useCareerFormActions';
-export { useCreateFormModal } from './useCreateFormModal';
