+++ src/features/career-form/ui/create-form-modal/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useCreateFormModal } from './useCreateFormModal';
