+++ src/features/career-form/ui/create-form-modal/styles.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+import { Modal } from '@admiral-ds/react-ui';
+
+export const StyledModal = styled(Modal)`
+  top: 35%;
+`;
